Implement IInteractable
