+++ Interfaces/IInteractable.cs
@@ -1,0 +1,5 @@
+
+public interface IInteractable
+{
+    public void OnInteract();
+}
